Changed to service
The initial intent was to create a command-line executable which could be scheduled with sc create.  But then I realized it could just be built as a service with an installer and everything.  This has the added benefit of allowing me to send commands to the child process(es) so everything shuts down cleanly.

Name changes were required.  ServiceLauncher doesn't make as much sense as LauncherService, given that it's a service that launches other programs.  The command line module, for clarity, was renamed to LauncherControl (because it's used to control LauncherService).